requirements.txt
django
djangorestframework
django-cors-headers
django-phonenumber-field
djangorestframework-simplejwt
python-dotenv
phonenumbers
psycopg2
Pillow
Faker
dotenv